--- .gulp.json
@@ -1,5 +1,0 @@
-{
-    "flags": {
-        "gulpfile": "gulpfile.cjs"
-    }
-}
