+++ package.json
@@ -73,14 +73,13 @@
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^22.0.2",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "aws-sdk": "^2.932.0",
-    "bluebird": "^3.7.2",
+    "@rollup/plugin-node-resolve": "^14.0.0",
+    "aws-sdk": "^2.1210.0",
     "chokidar": "^3.5.3",
     "cloc": "^2.8.0",
     "compression": "^1.7.4",
     "draco3d": "^1.5.1",
-    "esbuild": "^0.15.6",
+    "esbuild": "^0.15.7",
     "eslint": "^8.19.0",
     "eslint-config-cesium": "^9.0.0",
     "eslint-config-prettier": "^8.3.0",
@@ -88,7 +87,7 @@
     "eslint-plugin-html": "^6.1.2",
     "eslint-plugin-node": "^11.1.0",
     "express": "^4.17.1",
-    "globby": "^11.0.4",
+    "globby": "^13.1.2",
     "glsl-strip-comments": "^1.0.0",
     "gulp": "^4.0.2",
     "gulp-clean-css": "^4.3.0",
@@ -115,9 +114,11 @@
     "karma-spec-reporter": "^0.0.32",
     "markdownlint-cli": "^0.31.1",
     "merge-stream": "^2.0.0",
-    "mime": "^2.5.2",
+    "mime": "^3.0.0",
     "mkdirp": "^1.0.4",
+    "node-fetch": "^3.2.10",
     "open": "^8.2.1",
+    "p-limit": "^4.0.0",
     "prettier": "2.1.2",
     "prismjs": "^1.28.0",
     "request": "^2.79.0",
@@ -132,20 +133,20 @@
   },
   "scripts": {
     "prepare": "gulp prepare && husky install",
-    "start": "node server.cjs",
-    "start-public": "node server.cjs --public",
+    "start": "node server.js",
+    "start-public": "node server.js --public",
     "build": "gulp build",
-    "build-watch": "gulp build-watch",
-    "build-ts": "gulp build-ts",
-    "build-third-party": "gulp build-third-party",
-    "build-apps": "gulp build-apps",
+    "build-watch": "gulp buildWatch",
+    "build-ts": "gulp buildTs",
+    "build-third-party": "gulp buildThirdParty",
+    "build-apps": "gulp buildApps",
     "clean": "gulp clean",
     "cloc": "gulp cloc",
     "coverage": "gulp coverage",
-    "build-docs": "gulp build-docs",
-    "build-docs-watch": "gulp build-docs-watch",
+    "build-docs": "gulp buildDocs",
+    "build-docs-watch": "gulp buildDocsWatch",
     "eslint": "eslint \"./**/*.js\" \"./**/*.cjs\" \"./**/*.html\" --cache --quiet",
-    "make-zip": "gulp make-zip",
+    "make-zip": "gulp makeZip",
     "markdownlint": "markdownlint \"*.md\" \"Documentation/**/*.md\" --ignore CHANGES.md --ignore README.md --ignore LICENSE.md",
     "release": "gulp release",
     "test": "gulp test",
@@ -155,9 +156,9 @@
     "test-webgl-validation": "gulp test --webglValidation",
     "test-webgl-stub": "gulp test --webglStub",
     "test-release": "gulp test --release",
-    "deploy-s3": "gulp deploy-s3",
-    "deploy-status": "gulp deploy-status",
-    "deploy-set-version": "gulp deploy-set-version",
+    "deploy-s3": "gulp deployS3",
+    "deploy-status": "gulp deployStatus",
+    "deploy-set-version": "gulp deploySetVersion",
     "prettier": "prettier --write \"**/*\"",
     "prettier-check": "prettier --check \"**/*\""
   },
